Start render screens randomly
