+++ js/data/level-types.js
@@ -1,0 +1,11 @@
+export default [
+  {
+    type: `one-image`
+  },
+  {
+    type: `two-images`
+  },
+  {
+    type: `three-images`
+  }
+];
