+++ js/utils.js
@@ -10,13 +10,14 @@ export const createElementFromTemplate = (template) => {
 
   return container;
 };
-
-export const generateRandomNumber = (min, max) => Math.floor(Math.random() * (max - min + 1)) + min;
-
 /**
- * Get random image from data
- * @return {String}
+ * Generate whole random number with min and max inclusively
+ * @param {Number} min
+ * @param {Number} max
+ * @return {Number}
  */
-export const getRandomImage = (images) => {
-  return images[generateRandomNumber(0, images.length)];
+export const generateRandomNumber = (min, max) => Math.floor(Math.random() * (max - min) + min);
+
+export const getRandomElement = (array) => {
+  return array[generateRandomNumber(0, array.length)];
 };
